REF: moving flattenIO into a dedicated JSON Schema helpers file.
