+++ express-zod-api/src/documentation-helpers.ts
@@ -45,7 +45,8 @@ import { ezDateOutBrand } from "./date-out-schema";
 import { contentTypes } from "./content-type";
 import { DocumentationError } from "./errors";
 import { ezFileBrand } from "./file-schema";
-import { flattenIO, IOSchema } from "./io-schema";
+import { IOSchema } from "./io-schema";
+import { flattenIO } from "./json-schema-helpers";
 import { Alternatives } from "./logical-container";
 import { metaSymbol } from "./metadata";
 import { Method } from "./method";
