+++ express-zod-api/tests/io-schema.spec.ts
@@ -1,7 +1,7 @@
 import { expectTypeOf } from "vitest";
 import { z } from "zod";
 import { IOSchema, Middleware, ez } from "../src";
-import { flattenIO, getFinalEndpointInputSchema } from "../src/io-schema";
+import { getFinalEndpointInputSchema } from "../src/io-schema";
 import { metaSymbol } from "../src/metadata";
 import { AbstractMiddleware } from "../src/middleware";
 
@@ -283,77 +283,6 @@ describe("I/O Schema and related helpers", () => {
           five: "some",
         },
       ]);
-    });
-  });
-
-  describe("flattenIO()", () => {
-    test("should pass the object schema through", () => {
-      const subject = flattenIO({
-        type: "object",
-        properties: { one: { type: "string" } },
-        required: ["one"],
-        examples: [{ one: "test" }],
-      });
-      expect(subject).toMatchSnapshot();
-    });
-
-    test("should return object schema for the union of object schemas", () => {
-      const subject = flattenIO({
-        oneOf: [
-          {
-            type: "object",
-            properties: { one: { type: "string" } },
-            required: ["one"],
-            examples: [{ one: "test" }],
-          },
-          {
-            type: "object",
-            properties: { two: { type: "number" } },
-            required: ["two"],
-            examples: [{ two: "jest" }],
-          },
-        ],
-      });
-      expect(subject).toMatchSnapshot();
-    });
-
-    test("should return object schema for the intersection of object schemas", () => {
-      const subject = flattenIO({
-        allOf: [
-          {
-            type: "object",
-            properties: { one: { type: "string" } },
-            required: ["one"],
-            examples: [{ one: "test" }],
-          },
-          {
-            type: "object",
-            properties: { two: { type: "number" } },
-            required: ["two"],
-            examples: [{ two: "jest" }],
-          },
-        ],
-      });
-      expect(subject).toMatchSnapshot();
-    });
-
-    test("should handle records", () => {
-      const subject = z.toJSONSchema(
-        z
-          .record(z.literal(["one", "two"]), z.string())
-          .meta({
-            examples: [
-              { one: "test", two: "jest" },
-              { one: "some", two: "another" },
-            ],
-          })
-          .or(
-            z
-              .record(z.enum(["three", "four"]), z.number())
-              .meta({ examples: [{ three: 123, four: 456 }] }),
-          ),
-      );
-      expect(flattenIO(subject)).toMatchSnapshot();
     });
   });
 });
